MPT: Fix invalid config setting
We are getting:

```
WARN  2025-06-17 14:18:07,060 [shard 0:main] cluster - config_manager.cc:459 - Invalid bootstrap property 'topic_memory_per_partition' validation error: Invalid value: YAML::TypedBadConversion<unsigned long> (yaml-cpp: error at line 31, column 29: bad conversion)
```

Which I suspect is because we pass a float (XXX.0) instead of an int. This just
seems to get silently ignored but it doesn't make a functional difference as we
are just setting the default value anywa.

+++ tests/rptest/scale_tests/many_partitions_test.py
@@ -770,7 +770,7 @@ class ManyPartitionsTest(PreallocNodesTest):
             'topic_partitions_per_shard':
             DEFAULT_PARTITIONS_PER_SHARD,
             'topic_memory_per_partition':
-            DEFAULT_MIB_PER_PARTITION * 1024 * 1024,
+            int(DEFAULT_MIB_PER_PARTITION * 1024 * 1024),
             'topic_partitions_memory_allocation_percent':
             DEFAULT_PARTITIONS_MEMORY_ALLOCATION_PERCENT,
         })
@@ -860,7 +860,7 @@ class ManyPartitionsTest(PreallocNodesTest):
             'topic_partitions_per_shard':
             topic_partitions_per_shard,
             'topic_memory_per_partition':
-            mib_per_partition * 1024 * 1024,
+            int(mib_per_partition * 1024 * 1024),
             'topic_partitions_memory_allocation_percent':
             DEFAULT_PARTITIONS_MEMORY_ALLOCATION_PERCENT,
         })
